Moving funtions from util to lium_paths
