+++ src/config.rs
@@ -16,7 +16,7 @@ use regex::Regex;
 use serde::Deserialize;
 use serde::Serialize;
 
-use crate::util::gen_path_in_lium_dir;
+use crate::util::lium_paths::gen_path_in_lium_dir;
 use crate::util::shell_helpers::run_bash_command;
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
